requirements.txt
flask
flask-restx
flask_jwt_extended
werkzeug==0.16.1
authlib==0.15.3
flask-injector==0.12.3
attrs==20.3.0
cattrs==1.3.0
markupsafe==1.1.1
psycopg2
Flask-Migrate
Flask-SQLAlchemy
psycopg2-binary
requests
pytest
pydantic
sendgrid
boto3


